fix compile error

+++ src/source/Ice/IceAgent.c
@@ -1820,7 +1820,7 @@ STATUS iceAgentInitRelayCandidate(PIceAgent pIceAgent, UINT32 iceServerIndex, KV
     pNewCandidate->priority = computeCandidatePriority(pNewCandidate);
 
     TurnConnectionCallbacks callback = {0};
-    callback.customData = pNewCandidate;
+    callback.customData = (UINT64)pNewCandidate;
     callback.relayAddressAvailableFn = NULL;
     callback.turnStateFailedFn= trunStateFailedFn;
